just renamed class
