+++ src/main/java/uk/gov/hmcts/ccd/endpoint/DocLinksRestoreEndpoint.java
@@ -7,7 +7,7 @@ import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 import uk.gov.hmcts.ccd.data.casedetails.CaseDetailsEntity;
-import uk.gov.hmcts.ccd.domain.service.DocLinksRestoreService;
+import uk.gov.hmcts.ccd.domain.service.doclink.DocLinksDetectionService;
 
 import java.util.ArrayList;
 import java.util.Arrays;
@@ -17,11 +17,11 @@ import java.util.stream.Collectors;
 @RestController
 public class DocLinksRestoreEndpoint {
 
-    private final DocLinksRestoreService docLinksRestoreService;
+    private final DocLinksDetectionService docLinksDetectionService;
 
     @Autowired
-    public DocLinksRestoreEndpoint(DocLinksRestoreService docLinksRestoreService) {
-        this.docLinksRestoreService = docLinksRestoreService;
+    public DocLinksRestoreEndpoint(DocLinksDetectionService docLinksDetectionService) {
+        this.docLinksDetectionService = docLinksDetectionService;
     }
 
     @RequestMapping(value = "/doclinks/restore", method = RequestMethod.GET)
@@ -29,7 +29,7 @@ public class DocLinksRestoreEndpoint {
         @RequestParam(value = "jids", required = false) final String jurisdictionIds) {
         List<String> jurisdictionList = StringUtils.isNotEmpty(jurisdictionIds) ?
             Arrays.asList(jurisdictionIds.split(",")) : new ArrayList<>();
-        List<CaseDetailsEntity> docLinksMissedCases = docLinksRestoreService.findDocLinksMissedCases(jurisdictionList);
+        List<CaseDetailsEntity> docLinksMissedCases = docLinksDetectionService.findDocLinksMissedCases(jurisdictionList);
         return docLinksMissedCases.stream().map(c -> c.getId()).collect(Collectors.toList());
     }
 }
